- Move layout constants to CLinker interface - Move platform specific constants to internal class
