+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/CABI.java
@@ -4,7 +4,9 @@
  *
  *  This code is free software; you can redistribute it and/or modify it
  *  under the terms of the GNU General Public License version 2 only, as
- *  published by the Free Software Foundation.
+ *  published by the Free Software Foundation.  Oracle designates this
+ *  particular file as subject to the "Classpath" exception as provided
+ *  by Oracle in the LICENSE file that accompanied this code.
  *
  *  This code is distributed in the hope that it will be useful, but WITHOUT
  *  ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
@@ -16,29 +18,30 @@
  *  2 along with this work; if not, write to the Free Software Foundation,
  *  Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  *
- *  Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ *   Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
  *  or visit www.oracle.com if you need additional information or have any
  *  questions.
+ *
  */
+package jdk.internal.foreign;
 
-/*
- * @test
- * @run testng/othervm TestCircularInit2
- */
+public enum CABI {
+    SysV,
+    Win64,
+    AArch64;
 
-import jdk.incubator.foreign.CLinker;
-import org.testng.annotations.Test;
-
-import static org.testng.Assert.assertNotNull;
-
-public class TestCircularInit2 {
-
-    @Test
-    public void testCircularInit() {
-        System.out.println(CLinker.C_BOOL); // trigger clinit
-        assertNotNull(CLinker.C_BOOL);
-        assertNotNull(CLinker.C_BOOL);
-        assertNotNull(CLinker.C_BOOL);
+    public static CABI current() {
+        String arch = System.getProperty("os.arch");
+        String os = System.getProperty("os.name");
+        if (arch.equals("amd64") || arch.equals("x86_64")) {
+            if (os.startsWith("Windows")) {
+                return Win64;
+            } else {
+                return SysV;
+            }
+        } else if (arch.equals("aarch64")) {
+            return AArch64;
+        }
+        throw new UnsupportedOperationException("Unsupported os or arch: " + os + ", " + arch);
     }
-
 }
