+++ test/jdk/java/foreign/StdLibTest.java
@@ -23,6 +23,7 @@
 
 /*
  * @test
+ * @modules jdk.incubator.foreign/jdk.internal.foreign
  * @build NativeTestHelper StdLibTest
  * @run testng/othervm -Dforeign.restricted=permit StdLibTest
  */
@@ -50,13 +51,13 @@ import static jdk.incubator.foreign.MemoryAccess.*;
 
 import org.testng.annotations.*;
 
-import static jdk.incubator.foreign.CSupport.*;
+import static jdk.incubator.foreign.CLinker.*;
 import static org.testng.Assert.*;
 
 @Test
 public class StdLibTest extends NativeTestHelper {
 
-    final static ForeignLinker abi = CSupport.getSystemLinker();
+    final static ForeignLinker abi = CLinker.getSystemLinker();
 
     private StdLibHelper stdLibHelper = new StdLibHelper();
