+++ test/jdk/java/foreign/callarranger/TestWindowsCallArranger.java
@@ -24,7 +24,8 @@
 
 /*
  * @test
- * @modules jdk.incubator.foreign/jdk.internal.foreign.abi
+ * @modules jdk.incubator.foreign/jdk.internal.foreign
+ *          jdk.incubator.foreign/jdk.internal.foreign.abi
  *          jdk.incubator.foreign/jdk.internal.foreign.abi.x64
  *          jdk.incubator.foreign/jdk.internal.foreign.abi.x64.windows
  * @build CallArrangerTestBase
@@ -42,7 +43,7 @@ import org.testng.annotations.Test;
 
 import java.lang.invoke.MethodType;
 
-import static jdk.incubator.foreign.CSupport.Win64.*;
+import static jdk.internal.foreign.PlatformLayouts.Win64.*;
 import static jdk.internal.foreign.abi.Binding.*;
 import static jdk.internal.foreign.abi.Binding.copy;
 import static jdk.internal.foreign.abi.x64.X86_64Architecture.*;
